Fix floating failing rake tasks spec
This spec was intended to check, that calling process_resend! for
controller, that was not accidentally inherited from ActionController
and not having receivers_by_action method - won't fail and will just
return nil. So, the test class was inherited from ChannelController.
But there is a problem - if this test class is inherited from ChannelController,
it leads to channels_spec failing.
The GC cannot remove test class in time and it exists while channels_spec
executing sometimes`. There are extra descendants of ChannelController
and test goes wrong. So I changed spec to simpy remove desired method.